task_manager_internship-main/requirements.txt
Django>=5.2,<6
daphne>=4.1,<5
djangorestframework>=3.15,<4
mysqlclient>=2.2
deep-translator>=1.11,<2
langdetect>=1.0,<2
channels>=4.1,<5
channels-redis>=4.2,<5
